The methods that are common to UL and DL are moved to the Simulation abstract base class.
